split agent and rpc
